Initial UWP adaptation
(cherry picked from commit 87bdef086f64cc8f301d4f4b7f294f6b0fbdb436)

# Conflicts:
#	.gitignore
#	CMakeLists.txt
#	src/CMakeLists.txt
#	src/sdl/CMakeLists.txt
#	vcpkg.json

+++ src/m_easing.c
@@ -87,7 +87,7 @@ static fixed_t fixlog2(fixed_t a)
 	}
 
 	if (a <= FRACUNIT)
-		return -y;
+		return -(INT32)y;
 
 	return y;
 }
